Move CEC Frustration tests into Card_4_142_Tests.
